+++ migrations/20250309112219_add_date_of_birth_to_users.js
@@ -1,0 +1,19 @@
+/**
+ * @param { import("knex").Knex } knex
+ * @returns { Promise<void> }
+ */
+exports.up = function (knex) {
+  return knex.schema.withSchema('users').table('user', function (table) {
+    table.timestamp('date_of_birth');
+  });
+};
+
+/**
+ * @param { import("knex").Knex } knex
+ * @returns { Promise<void> }
+ */
+exports.down = function (knex) {
+  return knex.schema.withSchema('users').table('user', function (table) {
+    table.dropColumn('date_of_birth');
+  });
+};