+++ src/app/register/page.tsx
@@ -1,10 +1,9 @@
-import { Main } from '@/components/ui/Main';
 import { RegisterForm } from '@/features/users/register/components/RegisterForm';
 
 export default async function Register() {
   return (
-    <Main centered>
+    <main className='lg:px-default xs:px-4 flex flex-col gap-2 w-full flex-1 overflow-y-scroll py-4'>
       <RegisterForm />
-    </Main>
+    </main>
   );
 }