+++ src/features/users/register/hooks/useRegisterForm.tsx
@@ -1,14 +1,29 @@
 import { useRecord } from '@/hooks/useRecord';
 import { useStatus } from '@/hooks/useStatus';
-import { useCallback } from 'react';
+import { useCallback, useEffect, useMemo } from 'react';
 import { registerUserAction } from '../actions/registerUserAction';
 import toast from 'react-hot-toast';
 import { useRouter } from 'next/navigation';
 import { RegisterError } from '../types/RegisterError';
 import { registerCredentialsSchema } from '../schemas/registerCredentialsSchema';
+import { z } from 'zod';
+
+const dateOfBirthSchema = z
+  .string()
+  .date()
+  .refine(
+    date => {
+      const now = new Date();
+      const dob = new Date(date);
+      const age = now.getFullYear() - dob.getFullYear();
+      return age >= 18;
+    },
+    { message: 'You must be 18 years or older!' }
+  );
 
 export function useRegisterForm() {
   const { record: credentials, updateOnChange: updateCredentials } = useRecord({
+    dateOfBirth: '',
     email: '',
     password1: '',
     password2: '',
@@ -19,12 +34,13 @@ export function useRegisterForm() {
     'password_too_long',
     'password_too_short',
     'user_exists',
+    'underage',
   ]);
 
   const onSubmit = useCallback(
     async (e: React.FormEvent) => {
+      console.log('onSubmit');
       e.preventDefault();
-
       let currentStatus: typeof status = 'loading';
       setStatus(currentStatus);
       try {
@@ -36,23 +52,24 @@ export function useRegisterForm() {
           const result = await registerUserAction(credentials);
           if (result.code === 0) {
             toast.success('Registration succeeded!');
-            router.replace('/login');
+            router.replace(`/register/verify?email=${credentials.email}`);
             currentStatus = 'done';
           } else {
-            if (result.code === RegisterError.DUPLICATE_USER) {
-              toast.error('A user with the provided email already exists!');
-              currentStatus = 'user_exists';
-            } else if (result.code === RegisterError.USER_COUNT) {
+            if (result.code === RegisterError.USER_COUNT) {
               toast.error(
                 'Unfortunately, we do not allow registration of any more users at this point.'
               );
               currentStatus = 'error';
+            } else if (result.code === RegisterError.DUPLICATE_USER) {
+              currentStatus = 'user_exists';
             } else {
-              toast.error('An unexpected error occured!');
               currentStatus = 'error';
             }
           }
         }
+      } catch (err) {
+        toast.error('An unexpected error occured!');
+        currentStatus = 'error';
       } finally {
         setStatus(currentStatus);
       }
@@ -60,5 +77,17 @@ export function useRegisterForm() {
     [setStatus, router, registerUserAction, credentials]
   );
 
-  return { credentials, updateCredentials, onSubmit, status };
+  const registerButtonDisabled = useMemo(() => {
+    return status === 'done' || status === 'loading';
+  }, [status]);
+
+  useEffect(() => {
+    if (dateOfBirthSchema.safeParse(credentials.dateOfBirth).success === false) {
+      setStatus('underage');
+    } else {
+      setStatus('idle');
+    }
+  }, [credentials.dateOfBirth, setStatus]);
+
+  return { credentials, updateCredentials, onSubmit, status, registerButtonDisabled };
 }