+++ src/features/users/register/schemas/registerCredentialsSchema.ts
@@ -4,15 +4,27 @@ import z from 'zod';
 export const registerCredentialsSchema = z
   .object({
     email: z.string().email().nonempty().trim(),
+    dateOfBirth: z.string(),
     password1: passwordSchema,
     password2: passwordSchema,
   })
   .strict()
   .refine(
     value => {
+      const { dateOfBirth } = value;
+      const age = new Date().getFullYear() - new Date(dateOfBirth).getFullYear(); // Updated to use date object
+      return age >= 18;
+    },
+    {
+      message: 'underage',
+    }
+  )
+  .refine(
+    value => {
       //Make sure the passwords are the same.
       const { password1, password2 } = value;
-      return password1 === password2;
+      const passwordsEqual = password1 === password2;
+      return passwordsEqual;
     },
     {
       message: 'password_mismatch',