+++ src/middleware.ts
@@ -1,13 +1,35 @@
 import { getToken } from 'next-auth/jwt';
 import { NextRequestWithAuth } from 'next-auth/middleware';
 import { NextResponse } from 'next/server';
+import { RateLimiter } from './utils/rateLimiter';
+
+const rateLimiter = new RateLimiter({
+  requestLimit: 10,
+  cooldownTime: 120000,
+  ttl: 60000,
+  logging: true,
+});
 
 export default async function middleware(req: NextRequestWithAuth) {
   const token = await getToken({ req });
   const url = req.nextUrl.pathname;
+
+  if (
+    (req.method === 'POST' && url === '/register') ||
+    url === '/api/public/users/resend_verification_email' ||
+    url === '/api/public/users/verify'
+  ) {
+    //Limit the number of times a user is allowed to hit these endpoints.
+
+    const res = await rateLimiter.limit(req);
+    if (res.status !== 200) {
+      return res;
+    }
+  }
+
   if (token) {
     //Disallow access to the front- login- or register pages once logged in.
-    if (url === '/' || url === '/login' || url === '/register') {
+    if (url === '/' || url.includes('/login') || url.includes('/register')) {
       return redirectTo('/auth/dashboard', req);
     }
   } else {
@@ -16,8 +38,6 @@ export default async function middleware(req: NextRequestWithAuth) {
       return redirectTo('/login', req);
     }
   }
-
-  //TODO: Apply rate limiting to the /register, and /auth/bets/create POST-routes (server actions).
 
   return NextResponse.next();
 }
@@ -29,5 +49,5 @@ function redirectTo(url: string, req: NextRequestWithAuth) {
 }
 
 export const config = {
-  matcher: ['/', '/auth/:path*', '/login', '/register'],
+  matcher: ['/', '/auth/:path*', '/login', '/register', '/api/:path*'],
 };